+++ internal/provider/resource_nxos_rest.go
@@ -179,8 +179,8 @@ func (r resourceRest) Delete(ctx context.Context, req tfsdk.DeleteResourceReques
 	if err != nil {
 		errCode := res.Get("imdata.0.error.attributes.code").Str
 		// Ignore errors of type "Cannot delete object"
-		if errCode != "107" {
-			resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Failed to update object, got error: %s", err))
+		if errCode != "1" && errCode != "107" {
+			resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Failed to delete object, got error: %s", err))
 			return
 		}
 	}