+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 - Add nxos_bgp_peer_address_family resource and data source
 - Add nxos_bgp_route_control resource and data source
+- Add nxos_nve_interface resource and data source
+- Add nxos_nve_vni_container resource and data source
+- Add nxos_nve_vni resource and data source
+- Add nxos_nve_vni_ingress_replication resource and data source
 
 ## 0.3.3